Classroom Mini-Project

+++ classes/views.py
@@ -2,7 +2,9 @@ from django.shortcuts import render, redirect
 from django.contrib import messages
 
 from .models import Classroom
-from .forms import ClassroomForm
+from .forms import *
+
+from django.contrib.auth import login, authenticate, logout
 
 def classroom_list(request):
 	classrooms = Classroom.objects.all()
@@ -22,10 +24,14 @@ def classroom_detail(request, classroom_id):
 
 def classroom_create(request):
 	form = ClassroomForm()
+	if not request.user.is_authenticated:
+		return redirect('signin')
 	if request.method == "POST":
 		form = ClassroomForm(request.POST, request.FILES or None)
 		if form.is_valid():
-			form.save()
+			classroom = form.save(commit=False)
+			classroom.teacher = request.user
+			classroom.save()
 			messages.success(request, "Successfully Created!")
 			return redirect('classroom-list')
 		print (form.errors)
@@ -38,6 +44,8 @@ def classroom_create(request):
 def classroom_update(request, classroom_id):
 	classroom = Classroom.objects.get(id=classroom_id)
 	form = ClassroomForm(instance=classroom)
+	if not request.user.is_authenticated:
+		return redirect('signin')
 	if request.method == "POST":
 		form = ClassroomForm(request.POST, request.FILES or None, instance=classroom)
 		if form.is_valid():
@@ -53,6 +61,95 @@ def classroom_update(request, classroom_id):
 
 
 def classroom_delete(request, classroom_id):
+	if not request.user.is_authenticated:
+		return redirect('signin')
 	Classroom.objects.get(id=classroom_id).delete()
 	messages.success(request, "Successfully Deleted!")
 	return redirect('classroom-list')
+
+def student_create(request, classroom_id):
+	form = StudentForm()
+	classroom= Classroom.objects.get(id=classroom_id)
+	if request.user != classroom.teacher:
+		return redirect('signin')
+	if request.method == "POST":
+		form = StudentForm(request.POST, request.FILES or None)
+		if form.is_valid():
+			student = form.save(commit=False)
+			student.classroom = classroom
+			student.save()
+			messages.success(request, "Successfully Created!")
+			return redirect('classroom-detail', classroom_id)
+		print (form.errors)
+	context = {
+	"form": form,
+	"classroom" : classroom,
+	}
+	return render(request, 'students_create.html', context)
+
+
+def student_update(request, student_id):
+	student = Student.objects.get(id=student_id)
+	form = StudentForm(instance=student)
+	if request.user != student.classroom.teacher:
+		return redirect('signin')
+	if request.method == "POST":
+		form = StudentForm(request.POST, request.FILES or None, instance=student)
+		if form.is_valid():
+			form.save()
+			messages.success(request, "Successfully Edited!")
+			return redirect('classroom-detail', classroom_id=student.classroom.id)
+		print (form.errors)
+	context = {
+	"form": form,
+	"student": student,
+	}
+	return render(request, 'student_update.html', context)
+
+
+def student_delete(request, student_id):
+	if request.user != student.classroom.teacher:
+		return redirect('signin')
+	student = Student.objects.get(id=student_id)
+	student.delete()
+	messages.success(request, "Successfully Deleted!")
+	return redirect('classroom-detail', classroom_id=student.classroom.id)
+
+def signup(request):
+	form = SignupForm()
+	if request.method == 'POST':
+		form = SignupForm(request.POST)
+		if form.is_valid():
+			user = form.save(commit=False)
+
+			user.set_password(user.password)
+			user.save()
+
+			login(request, user)
+			return redirect("classroom-list")
+	context = {
+		"form":form,
+	}
+	return render(request, 'signup.html', context)
+
+def signin(request):
+	form = SigninForm()
+	if request.method == 'POST':
+		form = SigninForm(request.POST)
+		if form.is_valid():
+
+			username = form.cleaned_data['username']
+			password = form.cleaned_data['password']
+
+			auth_user = authenticate(username=username, password=password)
+			if auth_user is not None:
+				login(request, auth_user)
+				return redirect('classroom-list')
+	context = {
+		"form":form
+	}
+	return render(request, 'signin.html', context)
+
+def signout(request):
+	logout(request)
+	return redirect("signin")
